separated functional components save/load are still in control panel, for now
